refactor: modularize EvalsTab.tsx into 7 focused modules
