+++ src/app/(dashboard)/dashboard/analytics/evals/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import EvalsTab from "../../usage/components/EvalsTab";
+import EvalsTab from "../../usage/components/evals";
 
 export default function AnalyticsEvalsPage() {
   return <EvalsTab />;
